elee2045/personal assignment code/assignment-3-nmahadevan12/requirements.txt
Flask==3.0.2
paho-mqtt==2.1.0
requests==2.31.0